+++ activesupport/lib/active_support/core_ext/hash/keys.rb
@@ -73,34 +73,26 @@ class Hash
 
   # Return a new hash with all keys converted by the block operation.
   # This includes the keys from the root hash and from all
-  # nested hashes.
+  # nested hashes and arrays.
   #
   #  hash = { person: { name: 'Rob', age: '28' } }
   #
   #  hash.deep_transform_keys{ |key| key.to_s.upcase }
   #  # => { "PERSON" => { "NAME" => "Rob", "AGE" => "28" } }
   def deep_transform_keys(&block)
-    result = {}
-    each do |key, value|
-      result[yield(key)] = value.is_a?(Hash) ? value.deep_transform_keys(&block) : value
-    end
-    result
+    _deep_transform_keys_in_object(self, &block)
   end
 
   # Destructively convert all keys by using the block operation.
   # This includes the keys from the root hash and from all
-  # nested hashes.
+  # nested hashes and arrays.
   def deep_transform_keys!(&block)
-    keys.each do |key|
-      value = delete(key)
-      self[yield(key)] = value.is_a?(Hash) ? value.deep_transform_keys!(&block) : value
-    end
-    self
+    _deep_transform_keys_in_object!(self, &block)
   end
 
   # Return a new hash with all keys converted to strings.
   # This includes the keys from the root hash and from all
-  # nested hashes.
+  # nested hashes and arrays.
   #
   #   hash = { person: { name: 'Rob', age: '28' } }
   #
@@ -112,14 +104,14 @@ class Hash
 
   # Destructively convert all keys to strings.
   # This includes the keys from the root hash and from all
-  # nested hashes.
+  # nested hashes and arrays.
   def deep_stringify_keys!
     deep_transform_keys!{ |key| key.to_s }
   end
 
   # Return a new hash with all keys converted to symbols, as long as
   # they respond to +to_sym+. This includes the keys from the root hash
-  # and from all nested hashes.
+  # and from all nested hashes and arrays.
   #
   #   hash = { 'person' => { 'name' => 'Rob', 'age' => '28' } }
   #
@@ -131,8 +123,38 @@ class Hash
 
   # Destructively convert all keys to symbols, as long as they respond
   # to +to_sym+. This includes the keys from the root hash and from all
-  # nested hashes.
+  # nested hashes and arrays.
   def deep_symbolize_keys!
     deep_transform_keys!{ |key| key.to_sym rescue key }
   end
+
+  private
+    # support methods for deep transforming nested hashes and arrays
+    def _deep_transform_keys_in_object(object, &block)
+      case object
+      when Hash
+        object.each_with_object({}) do |(key, value), result|
+          result[yield(key)] = _deep_transform_keys_in_object(value, &block)
+        end
+      when Array
+        object.map {|e| _deep_transform_keys_in_object(e, &block) }
+      else
+        object
+      end
+    end
+
+    def _deep_transform_keys_in_object!(object, &block)
+      case object
+      when Hash
+        object.keys.each do |key|
+          value = object.delete(key)
+          object[yield(key)] = _deep_transform_keys_in_object!(value, &block)
+        end
+        object
+      when Array
+        object.map! {|e| _deep_transform_keys_in_object!(e, &block)}
+      else
+        object
+      end
+    end
 end